refact(e2e): Create env files for consistency
